+++ CHANGELOG.rst
@@ -7,7 +7,10 @@ dev
 
 **Improvements**
 
-- Update borders: Set ``SI – HU`` and ``SI – HU`` as Flow-based
+- Update borders: 
+    - Set ``SI – HU`` and ``SI – HU`` as Flow-based
+- Increase rate limits.
+- Increase default requests timeout.
 
 **Bugfixes**