+++ CHANGELOG.rst
@@ -12,6 +12,9 @@ dev
 - Increase rate limits.
 - Increase default requests timeout.
 - Add option to include ``proxies`` in ``EnergyQuantified`` and ``RealtoConnection``
+- Add support for using periods installed capacity instead of values when
+  converting a ``PeriodSeries`` to a ``Timeseries`` or a ``DataFrame``.
+
 
 **Bugfixes**